requirements.txt
addict==2.4.0
aenum==3.1.15
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
alabaster==0.7.16
albumentations==1.4.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
austin-dist==3.6.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.0
bleach==6.1.0
bqplot==0.12.42
branca==0.6.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cogeo-mosaic==7.1.0
color-operations==0.1.1
coloredlogs==15.0.1
colour==0.1.5
comm==0.1.4
commonmark==0.9.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deepforest==1.3.3
defusedxml==0.7.1
docopt==0.6.2
docstring_parser==0.16
docutils==0.20.1
efficientnet_pytorch==0.7.1
einops==0.7.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
fiona==1.9.5
flatbuffers==24.3.25
folium==0.14.0
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.0
gdown==4.7.1
geojson==3.0.1
geopandas==0.14.0
groundingdino-py==0.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.21.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
igraph==0.11.4
imagecodecs==2024.1.1
imageio==2.34.0
imagesize==1.4.1
importlib_metadata==7.0.2
importlib_resources==6.3.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.26.0
ipyleaflet==0.17.4
ipython==8.12.3
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonargparse==4.27.6
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
leafmap==0.27.1
lightly==1.4.7
lightly-utils==0.0.2
lightning==2.2.1
lightning-utilities==0.11.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
morecantile==5.0.0
mpmath==1.3.0
multidict==6.0.5
multimethod==1.11.2
munch==4.0.0
nbclient==0.8.0
nbconvert==7.16.3
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.1
onnxsim==0.4.36
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pipdeptree==2.16.2
platformdirs==3.11.0
pretrainedmodels==0.7.4
progressbar2==4.4.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==5.26.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
pyquadkey2==0.2.2
pyshp==2.3.1
PySocks==1.7.1
pystac==1.9.0
pystac-client==0.7.5
python-box==7.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-utils==3.8.2
pytorch-lightning==2.2.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
rasterio==1.3.9
recommonmark==0.7.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
reverse_geocoder==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rio-tiler==6.2.4
rpds-py==0.10.6
Rtree==1.2.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
scooby==0.9.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
slidingwindow==0.0.14
sniffio==1.3.0
snowballstemmer==2.2.0
snuggs==1.4.7
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
supermorecado==0.1.2
supervision==0.6.0
sympy==1.12
tdqm==0.0.1
tensorboardX==2.6.2.2
terminado==0.17.1
texttable==1.7.0
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchgeo==0.5.2
torchmetrics==1.3.2
torchpack @ git+https://github.com/zhijian-liu/torchpack.git@3a5a9f7ac665444e1eb45942ee3f8fc7ffbd84e5
torchprofile==0.0.4
torchvision==0.17.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
traittypes==0.2.1
transformers==4.38.2
triton==2.2.0
types-python-dateutil==2.8.19.14
typeshed_client==2.5.1
typing_extensions==4.10.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.9
xmltodict==0.13.0
xyzservices==2023.10.1
yapf==0.40.2
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1
